% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concepts}
\alias{concepts}
\title{Descriptions of variables in Systema Globalis.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 534 rows and 8 columns.}
\source{
\url{https://github.com/open-numbers/ddf--gapminder--systema_globalis}
}
\usage{
concepts
}
\description{
Descriptions of variables in Systema Globalis.
}
\keyword{datasets}
