% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Different variables identifying countries in the Systema Globalis database.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 273 rows and 33 columns.}
\source{
\url{https://github.com/open-numbers/ddf--gapminder--systema_globalis}
}
\usage{
countries
}
\description{
Different variables identifying countries in the Systema Globalis database.
}
\keyword{datasets}
