% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{datapoints}
\alias{datapoints}
\title{Observations of Systema Globalis' variables by country and year.}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 79290 rows and 534 columns.}
\source{
\url{https://github.com/open-numbers/ddf--gapminder--systema_globalis}
}
\usage{
datapoints
}
\description{
Observations of Systema Globalis' variables by country and year.
}
\keyword{datasets}
